Review - Clean Up & UI Fix
